requirements.txt
streamlit
boto3
sentence-transformers
faiss-cpu
speechrecognition
gtts
pandas
numpy